Front-End-new/requirements.txt
# Core AI and ML libraries
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0

# Database
supabase>=2.0.0

# Web framework for interface
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Environment and configuration
python-dotenv>=1.0.0

# Async support
asyncio-mqtt>=0.16.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Text processing
tiktoken>=0.5.0

# HTTP requests
requests>=2.31.0
aiohttp>=3.8.0

# Logging and utilities
rich>=13.0.0
click>=8.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
